Updated 2021 Day 8 Part 2
I ended up doing a bunch of extra work that I didn't need to do in order
to figure out what the output was.

For some reason I thought I had to recreate the 7 segment mapping from
the input to get the output instead of just matching the sorted
characters from the 10 input signals.

This means instead of trying to find individual segments I needed to
find numbers.